fix: allow tempfiles to be overwritten even while opened.
On Windows, it's common for newly created files, like persisted
tempfiles, to be opened by some antivirus program immediately.

When one then tries to write the same file again, like by
another tempfile-swap, the operation would fail as the target
file was opened.

Now tempfiles are created such that they have linux semantics when
they are opened, which allows them to be overwritten while the
reader holds on to the underlying original file.